requirement.txt
Pillow
numpy
scikit-image
scikit-learn
matplotlib
tensorflow
keras
fastapi
streamlit
uvicorn